# PERF: Fix performance regression in isin for empty values