--- a/doc/source/whatsnew/v1.5.3.rst
+++ b/doc/source/whatsnew/v1.5.3.rst
@@ -13,7 +13,7 @@ including other versions of pandas.
 
 Fixed regressions
 ~~~~~~~~~~~~~~~~~
--
+- Fixed performance regression in :meth:`Series.isin` when ``values`` is empty (:issue:`49839`
 -
 
 .. ---------------------------------------------------------------------------